+++ DownloadManager/Plugin.DownloadManager.Android/DownloadFileImplementation.cs
@@ -16,6 +16,19 @@ namespace Plugin.DownloadManager
 
         public string Url { get; set; }
 
+        private string _destinationPathName;
+
+        public string DestinationPathName
+        {
+            get { return _destinationPathName; }
+            set
+            {
+                if (Equals(_destinationPathName, value)) return;
+                _destinationPathName = value;
+                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(DestinationPathName)));
+            }
+        }
+
         public IDictionary<string, string> Headers { get; }
 
         private DownloadFileStatus _status;