+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changes
 
+### 1.4.0
+
+  * Feature [#94](https://github.com/SimonSimCity/Xamarin-CrossDownloadManager/issues/94): Added DestinationPathName to IDownloadFile which returns the path of the local file
+  * Bug [#88](https://github.com/SimonSimCity/Xamarin-CrossDownloadManager/issues/88): [Android] TotalBytesExpected and TotalBytesWritten is converted as integers but should be float
+
 ### 1.3.6
 
   * Fixed bug [#77](https://github.com/SimonSimCity/Xamarin-CrossDownloadManager/issues/77): [Android] Download are cancelled with no reason